* For addons, support more than just directories in the addons directories.
Addons can now be specified to --addons with:

addons/ADDON
ADDON
some-directory-not-in-the-tree

If we can't find the path specified in the --addon arg, we will look in
the addons directory to see if it exists there.

addons/auth-sssd will work the same as just auth-sssd.

Directories outside of the PWD must be absolute because we don't do any
path or env var expansion.

* Also making sure we only match a multi-line dockerfile command when it
ends with a slash followed by some or no whitespace. WORKDIR is now a
dockerfile command that we copy into our dockerfile for each container.

* auth-sssd Dockerfile no longer does a rm -rf * which would delete the
contents of your container and fail the build if no WORKDIR was set
(which is now fixed).