A random port was busting layer caching, so we're getting rid of it.
Default to port 8080 but let the user override it.

Multiple concurrent runs of the build.sh script will not have port
conflicts because they are running in docker, but running the go code
directly will cause ports to clash if running more than one instance.

We are also passing in sas-container-recipes-builder:PORT as an extra
host to be added to the image build /etc/hosts file so even if the IP of
the build container is changed, the hostname is consistent and will not
change the generated Dockerfile PLAYBOOK_SRV build arg (unless port is
changed between builds).